DAOS-5123 REBUILD: Rebuild Protocol and Tests for OSA Drain.
This patch includes the changes necessary for rebuild to migrate data
for OSA Drain.

Several modifications for rebuild to facilitate the drain
command/operation.
Refactored the rebuild tests for easier extensibility.
Implemented simple tests for Drain.
