+++ src/tests/suite/daos_rebuild.c
@@ -34,332 +34,12 @@
 #include <daos/mgmt.h>
 #include <daos/container.h>
 
-static test_arg_t *save_arg;
-
 #define KEY_NR		100
 #define OBJ_NR		10
 #define OBJ_CLS		OC_RP_3G1
 #define OBJ_REPLICAS	3
 #define DEFAULT_FAIL_TGT 0
 #define REBUILD_POOL_SIZE	(4ULL << 30)
-#define REBUILD_SUBTEST_POOL_SIZE (1ULL << 30)
-#define REBUILD_SMALL_POOL_SIZE (1ULL << 28)
-static void
-rebuild_exclude_tgt(test_arg_t **args, int arg_cnt, d_rank_t rank,
-		    int tgt_idx, bool kill)
-{
-	int i;
-
-	if (kill) {
-		daos_kill_server(args[0], args[0]->pool.pool_uuid,
-				 args[0]->group, &args[0]->pool.alive_svc,
-				 rank);
-		sleep(5);
-		/* If one rank is killed, then it has to exclude all
-		 * targets on this rank.
-		 **/
-		D_ASSERT(tgt_idx == -1);
-		return;
-	}
-
-	for (i = 0; i < arg_cnt; i++) {
-		daos_exclude_target(args[i]->pool.pool_uuid,
-				    args[i]->group, &args[i]->pool.svc,
-				    rank, tgt_idx);
-		sleep(2);
-	}
-}
-
-static void
-rebuild_add_tgt(test_arg_t **args, int args_cnt, d_rank_t rank,
-		int tgt_idx)
-{
-	int i;
-
-	for (i = 0; i < args_cnt; i++) {
-		if (!args[i]->pool.destroyed)
-			daos_add_target(args[i]->pool.pool_uuid,
-					args[i]->group,
-					&args[i]->pool.svc,
-					rank, tgt_idx);
-		sleep(2);
-	}
-}
-
-static void
-rebuild_targets(test_arg_t **args, int args_cnt, d_rank_t *ranks,
-		int *tgts, int rank_nr, bool fail, bool kill)
-{
-	int i;
-
-	for (i = 0; i < args_cnt; i++)
-		if (args[i]->rebuild_pre_cb)
-			args[i]->rebuild_pre_cb(args[i]);
-
-	MPI_Barrier(MPI_COMM_WORLD);
-	/** include or exclude the target from the pool */
-	if (args[0]->myrank == 0) {
-		for (i = 0; i < rank_nr; i++) {
-			if (fail) {
-				rebuild_exclude_tgt(args, args_cnt,
-						    ranks[i],
-						    tgts ? tgts[i] : -1,
-						    kill);
-			} else {
-				rebuild_add_tgt(args, args_cnt, ranks[i],
-						tgts ? tgts[i] : -1);
-
-			}
-			/* Sleep 5 seconds to make sure the rebuild start */
-			sleep(5);
-		}
-	}
-	MPI_Barrier(MPI_COMM_WORLD);
-
-	for (i = 0; i < args_cnt; i++)
-		if (args[i]->rebuild_cb)
-			args[i]->rebuild_cb(args[i]);
-
-	if (args[0]->myrank == 0)
-		test_rebuild_wait(args, args_cnt);
-
-	MPI_Barrier(MPI_COMM_WORLD);
-	for (i = 0; i < args_cnt; i++)
-		if (args[i]->rebuild_post_cb)
-			args[i]->rebuild_post_cb(args[i]);
-}
-
-static void
-rebuild_single_pool_rank(test_arg_t *arg, d_rank_t failed_rank)
-{
-	rebuild_targets(&arg, 1, &failed_rank, NULL, 1, true, false);
-}
-
-static void
-rebuild_pools_ranks(test_arg_t **args, int args_cnt, d_rank_t *failed_ranks,
-		    int ranks_nr)
-{
-	rebuild_targets(args, args_cnt, failed_ranks, NULL, ranks_nr, true,
-			false);
-}
-
-void
-rebuild_single_pool_target(test_arg_t *arg, d_rank_t failed_rank,
-			   int failed_tgt)
-{
-	rebuild_targets(&arg, 1, &failed_rank, &failed_tgt, 1, true, false);
-}
-
-void
-rebuild_add_back_tgts(test_arg_t *arg, d_rank_t failed_rank, int *failed_tgts,
-		      int nr)
-{
-	MPI_Barrier(MPI_COMM_WORLD);
-	/* Add back the target if it is not being killed */
-	if (arg->myrank == 0 && !arg->pool.destroyed) {
-		int i;
-
-		for (i = 0; i < nr; i++)
-			daos_add_target(arg->pool.pool_uuid, arg->group,
-					&arg->pool.svc, failed_rank,
-					failed_tgts ? failed_tgts[i] : -1);
-	}
-	MPI_Barrier(MPI_COMM_WORLD);
-}
-
-static int
-rebuild_io_obj_internal(struct ioreq *req, bool validate, daos_epoch_t eph,
-			daos_epoch_t validate_eph, int index)
-{
-#define BULK_SIZE	5000
-#define REC_SIZE	64
-#define LARGE_KEY_SIZE	(512 * 1024)
-#define DKEY_LOOP	3
-#define AKEY_LOOP	3
-#define REC_LOOP	10
-	char	dkey[32];
-	char	akey[32];
-	char	data[REC_SIZE];
-	char	data_verify[REC_SIZE];
-	char	*large_key;
-	int	akey_punch_idx = 1;
-	int	dkey_punch_idx = 1;
-	int	rec_punch_idx = 2;
-	int	j;
-	int	k;
-	int	l;
-
-	D_ALLOC(large_key, LARGE_KEY_SIZE);
-	if (large_key == NULL)
-		return -DER_NOMEM;
-	memset(large_key, 'L', LARGE_KEY_SIZE - 1);
-
-	for (j = 0; j < DKEY_LOOP; j++) {
-		req->iod_type = DAOS_IOD_ARRAY;
-		/* small records */
-		sprintf(dkey, "dkey_%d_%d", index, j);
-		sprintf(data, "%s_"DF_U64, "data", eph);
-		sprintf(data_verify, "%s_"DF_U64, "data", validate_eph);
-		for (k = 0; k < AKEY_LOOP; k++) {
-			sprintf(akey, "akey_%d_%d", index, k);
-			for (l = 0; l < REC_LOOP; l++) {
-				if (validate) {
-					/* How to verify punch? XXX */
-					if (k == akey_punch_idx ||
-					    j == dkey_punch_idx ||
-					    l == rec_punch_idx)
-						continue;
-					memset(data, 0, REC_SIZE);
-					if (l == 7)
-						lookup_single(large_key, akey,
-							      l, data, REC_SIZE,
-							      DAOS_TX_NONE,
-							      req);
-					else
-						lookup_single(dkey, akey, l,
-							      data, REC_SIZE,
-							      DAOS_TX_NONE,
-							      req);
-					assert_memory_equal(data, data_verify,
-						    strlen(data_verify));
-				} else {
-					if (l == 7)
-						insert_single(large_key, akey,
-							l, data,
-							strlen(data) + 1,
-							DAOS_TX_NONE, req);
-					else if (l == rec_punch_idx)
-						punch_single(dkey, akey, l,
-							     DAOS_TX_NONE,
-							     req);
-					else
-						insert_single(dkey, akey, l,
-							data, strlen(data) + 1,
-							DAOS_TX_NONE, req);
-				}
-			}
-
-			/* Punch akey */
-			if (k == akey_punch_idx && !validate)
-				punch_akey(dkey, akey, DAOS_TX_NONE, req);
-		}
-
-		/* large records */
-		for (k = 0; k < 2; k++) {
-			char bulk[BULK_SIZE+10];
-			char compare[BULK_SIZE];
-
-			sprintf(akey, "akey_bulk_%d_%d", index, k);
-			memset(compare, 'a', BULK_SIZE);
-			for (l = 0; l < 5; l++) {
-				if (validate) {
-					/* How to verify punch? XXX */
-					if (k == akey_punch_idx ||
-					    j == dkey_punch_idx)
-						continue;
-					memset(bulk, 0, BULK_SIZE);
-					lookup_single(dkey, akey, l,
-						      bulk, BULK_SIZE + 10,
-						      DAOS_TX_NONE, req);
-					assert_memory_equal(bulk, compare,
-							    BULK_SIZE);
-				} else {
-					memset(bulk, 'a', BULK_SIZE);
-					insert_single(dkey, akey, l,
-						      bulk, BULK_SIZE,
-						      DAOS_TX_NONE, req);
-				}
-			}
-
-			/* Punch akey */
-			if (k == akey_punch_idx && !validate)
-				punch_akey(dkey, akey, DAOS_TX_NONE, req);
-		}
-
-		/* Punch dkey */
-		if (j == dkey_punch_idx && !validate)
-			punch_dkey(dkey, DAOS_TX_NONE, req);
-
-		/* single record */
-		sprintf(data, "%s_"DF_U64, "single_data", eph);
-		sprintf(data_verify, "%s_"DF_U64, "single_data",
-			validate_eph);
-		req->iod_type = DAOS_IOD_SINGLE;
-		sprintf(dkey, "dkey_single_%d_%d", index, j);
-		if (validate) {
-			memset(data, 0, REC_SIZE);
-			lookup_single(dkey, "akey_single", 0, data, REC_SIZE,
-				      DAOS_TX_NONE, req);
-			assert_memory_equal(data, data_verify,
-					    strlen(data_verify));
-		} else {
-			insert_single(dkey, "akey_single", 0, data,
-				      strlen(data) + 1, DAOS_TX_NONE, req);
-		}
-	}
-
-	D_FREE(large_key);
-	return 0;
-}
-
-void
-rebuild_io(test_arg_t *arg, daos_obj_id_t *oids, int oids_nr)
-{
-	struct ioreq	req;
-	daos_epoch_t	eph = arg->hce + arg->index * 2 + 1;
-	int		i;
-	int		punch_idx = 1;
-
-	print_message("update obj %d eph "DF_U64" before rebuild\n", oids_nr,
-		      eph);
-
-	for (i = 0; i < oids_nr; i++) {
-		ioreq_init(&req, arg->coh, oids[i], DAOS_IOD_ARRAY, arg);
-		if (i == punch_idx) {
-			punch_obj(DAOS_TX_NONE, &req);
-		} else {
-			rebuild_io_obj_internal((&req), false, eph, -1,
-						 arg->index);
-		}
-		ioreq_fini(&req);
-	}
-}
-
-void
-rebuild_io_validate(test_arg_t *arg, daos_obj_id_t *oids, int oids_nr,
-		    bool discard)
-{
-	int	rc;
-	int	i;
-
-	for (i = 0; i < oids_nr; i++) {
-		/* XXX: skip punch object. */
-		if (i == 1)
-			continue;
-
-		rc = daos_obj_verify(arg->coh, oids[i], DAOS_EPOCH_MAX);
-		assert_int_equal(rc, 0);
-	}
-}
-
-/* Create a new pool for the sub_test */
-static int
-rebuild_pool_create(test_arg_t **new_arg, test_arg_t *old_arg, int flag,
-		    struct test_pool *pool)
-{
-	int rc;
-
-	/* create/connect another pool */
-	rc = test_setup((void **)new_arg, flag, old_arg->multi_rank,
-			REBUILD_SUBTEST_POOL_SIZE, pool);
-	if (rc) {
-		print_message("open/connect another pool failed: rc %d\n", rc);
-		return rc;
-	}
-
-	(*new_arg)->index = old_arg->index;
-	return 0;
-}
 
 /* Destroy the pool for the sub test */
 static void
@@ -370,157 +50,6 @@ rebuild_pool_destroy(test_arg_t *arg)
 	 * otherwise rebuild test might run into ENOSPACE
 	 */
 	sleep(1);
-}
-
-static int
-rebuild_pool_connect_internal(void *data)
-{
-	test_arg_t	*arg = data;
-	int		rc = 0;
-
-	MPI_Barrier(MPI_COMM_WORLD);
-	if (arg->myrank == 0) {
-		rc = daos_pool_connect(arg->pool.pool_uuid, arg->group,
-				       &arg->pool.svc, DAOS_PC_RW,
-				       &arg->pool.poh, &arg->pool.pool_info,
-				       NULL /* ev */);
-		if (rc)
-			print_message("daos_pool_connect failed, rc: %d\n", rc);
-
-		print_message("pool connect "DF_UUIDF"\n",
-			       DP_UUID(arg->pool.pool_uuid));
-	}
-	MPI_Barrier(MPI_COMM_WORLD);
-	if (arg->multi_rank)
-		MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
-	if (rc)
-		return rc;
-
-	/** broadcast pool info */
-	if (arg->multi_rank) {
-		MPI_Bcast(&arg->pool.pool_info, sizeof(arg->pool.pool_info),
-			  MPI_CHAR, 0, MPI_COMM_WORLD);
-		handle_share(&arg->pool.poh, HANDLE_POOL, arg->myrank,
-			     arg->pool.poh, 0);
-	}
-
-	/** open container */
-	MPI_Barrier(MPI_COMM_WORLD);
-	if (arg->myrank == 0) {
-		rc = daos_cont_open(arg->pool.poh, arg->co_uuid, DAOS_COO_RW,
-				    &arg->coh, &arg->co_info, NULL);
-		if (rc)
-			print_message("daos_cont_open failed, rc: %d\n", rc);
-
-		print_message("container open "DF_UUIDF"\n",
-			       DP_UUID(arg->co_uuid));
-	}
-	MPI_Barrier(MPI_COMM_WORLD);
-	if (arg->multi_rank)
-		MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
-	if (rc)
-		return rc;
-
-	/** broadcast container info */
-	if (arg->multi_rank) {
-		MPI_Bcast(&rc, 1, MPI_INT, 0, MPI_COMM_WORLD);
-		handle_share(&arg->coh, HANDLE_CO, arg->myrank, arg->pool.poh,
-			     0);
-	}
-
-	return 0;
-}
-
-static int
-rebuild_pool_disconnect_internal(void *data)
-{
-	test_arg_t	*arg = data;
-	int		rc = 0;
-	int		rc_reduce = 0;
-
-	/* Close cont and disconnect pool */
-	rc = daos_cont_close(arg->coh, NULL);
-	if (arg->multi_rank) {
-		MPI_Allreduce(&rc, &rc_reduce, 1, MPI_INT, MPI_MIN,
-			      MPI_COMM_WORLD);
-		rc = rc_reduce;
-	}
-	print_message("container close "DF_UUIDF"\n",
-		      DP_UUID(arg->co_uuid));
-	if (rc) {
-		print_message("failed to close container "DF_UUIDF
-			      ": %d\n", DP_UUID(arg->co_uuid), rc);
-		return rc;
-	}
-
-	arg->coh = DAOS_HDL_INVAL;
-	rc = daos_pool_disconnect(arg->pool.poh, NULL /* ev */);
-	if (rc)
-		print_message("failed to disconnect pool "DF_UUIDF
-			      ": %d\n", DP_UUID(arg->pool.pool_uuid), rc);
-
-	print_message("pool disconnect "DF_UUIDF"\n",
-		      DP_UUID(arg->pool.pool_uuid));
-
-	arg->pool.poh = DAOS_HDL_INVAL;
-	MPI_Barrier(MPI_COMM_WORLD);
-	return rc;
-}
-
-void
-reintegrate_single_pool_target(test_arg_t *arg, d_rank_t failed_rank,
-			       int failed_tgt)
-{
-	/* XXX: Disconnecting and reconnecting is necessary for the time being
-	 * while reintegration only supports "offline" mode and without
-	 * incremental reintegration. Disconnecting from the pool allows the
-	 * containers to be deleted before reintegration occurs
-	 *
-	 * Once incremental reintegration support is added, this should be
-	 * removed
-	 */
-	rebuild_pool_disconnect_internal(arg);
-	rebuild_targets(&arg, 1, &failed_rank, &failed_tgt, 1, false, false);
-	rebuild_pool_connect_internal(arg);
-}
-
-static void
-reintegrate_single_pool_rank(test_arg_t *arg, d_rank_t failed_rank)
-{
-
-	/* XXX: Disconnecting and reconnecting is necessary for the time being
-	 * while reintegration only supports "offline" mode and without
-	 * incremental reintegration. Disconnecting from the pool allows the
-	 * containers to be deleted before reintegration occurs
-	 *
-	 * Once incremental reintegration support is added, this should be
-	 * removed
-	 */
-	rebuild_pool_disconnect_internal(arg);
-	rebuild_targets(&arg, 1, &failed_rank, NULL, 1, false, false);
-	rebuild_pool_connect_internal(arg);
-}
-
-static void
-reintegrate_pools_ranks(test_arg_t **args, int args_cnt, d_rank_t *failed_ranks,
-			int ranks_nr)
-{
-	int i;
-
-	/* XXX: Disconnecting and reconnecting is necessary for the time being
-	 * while reintegration only supports "offline" mode and without
-	 * incremental reintegration. Disconnecting from the pool allows the
-	 * containers to be deleted before reintegration occurs
-	 *
-	 * Once incremental reintegration support is added, this should be
-	 * removed
-	 */
-	for (i = 0; i < args_cnt; i++)
-		rebuild_pool_disconnect_internal(args[i]);
-	rebuild_targets(args, args_cnt, failed_ranks, NULL, ranks_nr, false,
-			false);
-	for (i = 0; i < args_cnt; i++)
-		rebuild_pool_connect_internal(args[i]);
 }
 
 static void
@@ -549,7 +78,7 @@ rebuild_drop_scan(void **state)
 				     0, NULL);
 
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt);
+	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt, false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 
 	reintegrate_single_pool_target(arg, ranks_to_kill[0], tgt);
@@ -581,7 +110,7 @@ rebuild_retry_rebuild(void **state)
 				     DAOS_REBUILD_NO_HDL | DAOS_FAIL_ONCE,
 				     0, NULL);
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt);
+	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt, false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 
 	reintegrate_single_pool_target(arg, ranks_to_kill[0], tgt);
@@ -611,7 +140,7 @@ rebuild_retry_for_stale_pool(void **state)
 				     DAOS_REBUILD_STALE_POOL | DAOS_FAIL_ONCE,
 				     0, NULL);
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 
 	reintegrate_single_pool_rank(arg, ranks_to_kill[0]);
@@ -641,7 +170,7 @@ rebuild_drop_obj(void **state)
 				     DAOS_REBUILD_DROP_OBJ | DAOS_FAIL_ONCE,
 				     0, NULL);
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 
 	reintegrate_single_pool_rank(arg, ranks_to_kill[0]);
@@ -673,7 +202,7 @@ rebuild_update_failed(void **state)
 				     DAOS_REBUILD_UPDATE_FAIL | DAOS_FAIL_ONCE,
 				     0, NULL);
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt);
+	rebuild_single_pool_target(arg, ranks_to_kill[0], tgt, false);
 	reintegrate_single_pool_target(arg, ranks_to_kill[0], tgt);
 }
 
@@ -703,7 +232,7 @@ rebuild_multiple_pools(void **state)
 	rebuild_io(args[0], oids, OBJ_NR);
 	rebuild_io(args[1], oids, OBJ_NR);
 
-	rebuild_pools_ranks(args, 2, ranks_to_kill, 1);
+	rebuild_pools_ranks(args, 2, ranks_to_kill, 1, false);
 
 	rebuild_io_validate(args[0], oids, OBJ_NR, true);
 	rebuild_io_validate(args[1], oids, OBJ_NR, true);
@@ -803,7 +332,7 @@ rebuild_destroy_container(void **state)
 
 	new_arg->rebuild_cb = rebuild_destroy_container_cb;
 
-	rebuild_single_pool_rank(new_arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(new_arg, ranks_to_kill[0], false);
 
 	rebuild_pool_destroy(new_arg);
 }
@@ -834,7 +363,7 @@ rebuild_close_container(void **state)
 
 	new_arg->rebuild_pre_cb = rebuild_close_container_cb;
 
-	rebuild_single_pool_rank(new_arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(new_arg, ranks_to_kill[0], false);
 
 	rebuild_pool_destroy(new_arg);
 }
@@ -900,7 +429,7 @@ rebuild_destroy_pool_internal(void **state, uint64_t fail_loc)
 
 	new_arg->rebuild_cb = rebuild_destroy_pool_cb;
 
-	rebuild_single_pool_rank(new_arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(new_arg, ranks_to_kill[0], false);
 }
 
 static void
@@ -939,7 +468,7 @@ rebuild_iv_tgt_fail(void **state)
 				     DAOS_REBUILD_TGT_IV_UPDATE_FAIL |
 				     DAOS_FAIL_ONCE, 0, NULL);
 	MPI_Barrier(MPI_COMM_WORLD);
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 
 	reintegrate_single_pool_rank(arg, ranks_to_kill[0]);
@@ -973,7 +502,7 @@ rebuild_tgt_start_fail(void **state)
 	MPI_Barrier(MPI_COMM_WORLD);
 
 	/* Rebuild rank 1 */
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
 }
 
@@ -1002,7 +531,7 @@ rebuild_send_objects_fail(void **state)
 	/* Even do not sending the objects, the rebuild should still be
 	 * able to finish.
 	 */
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 
 	/* failed to start rebuild on rank 0 */
 	if (arg->myrank == 0)
@@ -1079,7 +608,7 @@ rebuild_tgt_pool_disconnect_internal(void **state, unsigned int fail_loc)
 	arg->rebuild_cb = rebuild_pool_disconnect_cb;
 	arg->rebuild_post_cb = rebuild_add_tgt_pool_connect_internal;
 
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 
 	arg->rebuild_cb = NULL;
 	arg->rebuild_post_cb = NULL;
@@ -1140,7 +669,7 @@ rebuild_offline_pool_connect_internal(void **state, unsigned int fail_loc)
 	arg->rebuild_pre_cb = rebuild_pool_disconnect_internal;
 	arg->rebuild_cb = rebuild_pool_connect_cb;
 
-	rebuild_targets(&arg, 1, ranks_to_kill, NULL, 1, true, true);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], true);
 
 	arg->rebuild_pre_cb = NULL;
 	arg->rebuild_cb = NULL;
@@ -1181,7 +710,7 @@ rebuild_offline(void **state)
 	arg->rebuild_pre_cb = rebuild_pool_disconnect_internal;
 	arg->rebuild_post_cb = rebuild_pool_connect_internal;
 
-	rebuild_targets(&arg, 1, ranks_to_kill, NULL, 1, true, true);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], true);
 
 	arg->rebuild_pre_cb = NULL;
 	arg->rebuild_post_cb = NULL;
@@ -1203,7 +732,7 @@ rebuild_offline_empty(void **state)
 	if (rc)
 		return;
 
-	rebuild_single_pool_rank(new_arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(new_arg, ranks_to_kill[0], false);
 	rebuild_pool_destroy(new_arg);
 
 }
@@ -1255,7 +784,7 @@ rebuild_master_change_during_scan(void **state)
 	MPI_Barrier(MPI_COMM_WORLD);
 	arg->rebuild_cb = rebuild_change_leader_cb;
 
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 
 	/* Verify the data */
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
@@ -1285,7 +814,7 @@ rebuild_master_change_during_rebuild(void **state)
 	MPI_Barrier(MPI_COMM_WORLD);
 	arg->rebuild_cb = rebuild_change_leader_cb;
 
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 
 	/* Verify the data */
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
@@ -1340,7 +869,7 @@ rebuild_nospace(void **state)
 	MPI_Barrier(MPI_COMM_WORLD);
 
 	arg->rebuild_cb = rebuild_nospace_cb;
-	rebuild_single_pool_rank(arg, ranks_to_kill[0]);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], false);
 
 	arg->rebuild_cb = NULL;
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
@@ -1467,7 +996,7 @@ rebuild_master_failure(void **state)
 	/* prepare the data */
 	rebuild_io(arg, oids, OBJ_NR);
 
-	rebuild_targets(&arg, 1, ranks_to_kill, NULL, 1, true, true);
+	rebuild_single_pool_rank(arg, ranks_to_kill[0], true);
 
 	/* Verify the data */
 	rebuild_io_validate(arg, oids, OBJ_NR, true);
@@ -1548,7 +1077,7 @@ rebuild_multiple_failures(void **state)
 #endif
 	arg->rebuild_post_cb_arg = cb_arg_oids;
 
-	rebuild_targets(&arg, 1, ranks_to_kill, NULL, MAX_KILLS, true, true);
+	rebuild_pools_ranks(&arg, 1, ranks_to_kill, MAX_KILLS, true);
 
 	arg->rebuild_cb = NULL;
 	arg->rebuild_post_cb = NULL;
@@ -1683,7 +1212,8 @@ multi_pools_rebuild_concurrently(void **state)
 	for (i = 0; i < POOL_NUM * CONT_PER_POOL; i++)
 		rebuild_io(args[i], oids, OBJ_PER_CONT);
 
-	rebuild_pools_ranks(args, POOL_NUM * CONT_PER_POOL, ranks_to_kill, 1);
+	rebuild_pools_ranks(args, POOL_NUM * CONT_PER_POOL, ranks_to_kill, 1,
+			false);
 
 	for (i = POOL_NUM * CONT_PER_POOL - 1; i >= 0; i--)
 		rebuild_io_validate(args[i], oids, OBJ_PER_CONT, true);
@@ -1691,51 +1221,6 @@ multi_pools_rebuild_concurrently(void **state)
 out:
 	for (i = POOL_NUM * CONT_PER_POOL - 1; i >= 0; i--)
 		rebuild_pool_destroy(args[i]);
-}
-
-static void
-save_group_state(void **state)
-{
-	if (state != NULL && *state != NULL) {
-		save_arg = *state;
-		*state = NULL;
-	}
-}
-
-int
-rebuild_sub_setup(void **state)
-{
-	save_group_state(state);
-	return test_setup(state, SETUP_CONT_CONNECT, true,
-			  REBUILD_SUBTEST_POOL_SIZE, NULL);
-}
-
-static void
-restore_group_state(void **state)
-{
-	if (state != NULL && save_arg != NULL) {
-		*state = save_arg;
-		save_arg = NULL;
-	}
-}
-
-int
-rebuild_sub_teardown(void **state)
-{
-	int rc;
-
-	rc = test_teardown(state);
-	restore_group_state(state);
-
-	return rc;
-}
-
-int
-rebuild_small_sub_setup(void **state)
-{
-	save_group_state(state);
-	return test_setup(state, SETUP_CONT_CONNECT, true,
-			  REBUILD_SMALL_POOL_SIZE, NULL);
 }
 
 /** create a new pool/container for each test */
