+++ src/tests/suite/daos_test.c
@@ -76,6 +76,7 @@ print_usage(int rank)
 	print_message("daos_test -O|--oid_alloc\n");
 	print_message("daos_test -r|--rebuild\n");
 	print_message("daos_test -v|--rebuild_simple\n");
+	print_message("daos_test -b|--drain_simple\n");
 	print_message("daos_test -N|--nvme_recovery\n");
 	print_message("daos_test -a|--daos_all_tests\n");
 	print_message("Default <daos_tests> runs all tests\n=============\n");
@@ -241,6 +242,13 @@ run_specified_tests(const char *tests, int rank, int size,
 			nr_failed += run_daos_rebuild_simple_test(rank, size,
 						sub_tests, sub_tests_size);
 			break;
+		case 'b':
+			daos_test_print(rank, "\n\n=================");
+			daos_test_print(rank, "DAOS drain simple tests..");
+			daos_test_print(rank, "=================");
+			nr_failed += run_daos_drain_simple_test(rank, size,
+						sub_tests, sub_tests_size);
+			break;
 
 		default:
 			D_ASSERT(0);
@@ -298,6 +306,7 @@ main(int argc, char **argv)
 		{"degraded",	no_argument,		NULL,	'd'},
 		{"rebuild",	no_argument,		NULL,	'r'},
 		{"rebuild_simple",	no_argument,	NULL,	'v'},
+		{"drain_simple",	no_argument,	NULL,	'b'},
 		{"nvme_recovery",	no_argument,	NULL,	'N'},
 		{"group",	required_argument,	NULL,	'g'},
 		{"csum_type",	required_argument,	NULL,
